+++ src/tests/sequences.nim
@@ -8,18 +8,20 @@ let intSeq = @[1,2,3,4,5,6,7,8,9,10]
 let floatSeq = @[1.1, 2.3, 3.3, 4.4, 5.5, 6.6, 7.7, 8.8, 9.9, 10.10]
 let boolSeq = @[true, true, false, false, false, true, true, true, true, false]
 
-proc dumpData() =
-    for (k,v) in QueryItr ^Sequence.kv:
-        echo k,"=",v
-
-proc dumpKeys() =
-    for key in QueryItr ^Sequence:
-        echo "key=", key
-
+const MAX_ELEMENTS = 1_000_000
+var hugeInt = newSeqOfCap[int](MAX_ELEMENTS) # seq[int]
+var hugeStr = newSeqOfCap[string](MAX_ELEMENTS) # seq[int]
+var hugeFloat = newSeqOfCap[float](MAX_ELEMENTS) # seq[int]
+var hugeBool = newSeqOfCap[bool](MAX_ELEMENTS) # seq[int]
+for i in 0..MAX_ELEMENTS:
+    hugeInt.add(i)
+    hugeStr.add($i)
+    hugeFloat.add(i.float * 1.25)
+    hugeBool.add(if i mod 3 == 0: true else: false)
 
 proc testStringSeq() =
     Set: ^Sequence(1) = join(strSeq, ",")
-    assert strSeq == Get ^Sequence(1).seqStr
+    assert strSeq == Get ^Sequence(1).seqString
 
 proc testIntSeq() =
     Set: ^Sequence(2) = join(intSeq, ",")
@@ -34,8 +36,9 @@ proc testBoolSeq() =
     assert boolSeq == Get ^Sequence(4).seqBool
 
 proc testSeq() =
+    Kill ^Sequence
     Set: ^Sequence(10) = strSeq
-    assert strSeq == Get ^Sequence(10).seqStr
+    assert strSeq == Get ^Sequence(10).seqString
     Set: ^Sequence(11) = intSeq
     assert intSeq == Get ^Sequence(11).seqInt
     Set: ^Sequence(12) = floatSeq
@@ -44,51 +47,73 @@ proc testSeq() =
     assert boolSeq == Get ^Sequence(13).seqBool
 
 proc testHugeSeq() =
-    var hugeInt: seq[int]
-    var hugeStr: seq[string]
-    var hugeFloat: seq[float]
-    var hugeBool: seq[bool]
-
-    for i in 0..500_000:
-        hugeInt.add(i)
-        hugeStr.add($i)
-        hugeFloat.add(i.float * 1.25)
-        hugeBool.add(if i mod 3 == 0: true else: false)
-
+    Kill ^Sequence
     timed:
         Set: ^Sequence("hugeStr") = hugeStr
-        assert hugeStr == Get ^Sequence("hugeStr").seqStr
-        echo hugeStr[0..10]
+        assert hugeStr == Get ^Sequence("hugeStr").seqString
 
     timed:
         Set: ^Sequence("hugeInt") = hugeInt
         assert hugeInt == Get ^Sequence("hugeInt").seqInt
-        echo hugeInt[0..10]
 
     timed:
         Set: ^Sequence("hugeFloat") = hugeFloat
         assert hugeFloat == Get ^Sequence("hugeFloat").seqFloat
-        echo hugeFloat[0..10]
 
     timed:
         Set: ^Sequence("hugeBool") = hugeBool
         assert hugeBool == Get ^Sequence("hugeBool").seqBool
-        echo hugeBool[0..10]
 
+
+proc testHugeSeqGzip() =
+    Kill ^Sequence
+    timed:
+        Set: ^Sequence("hugeStr") = hugeStr.gzip
+        assert hugeStr == Get ^Sequence("hugeStr").seqString.gzip
+
+    timed:
+        Set: ^Sequence("hugeInt") = hugeInt.gzip
+        assert hugeInt == Get ^Sequence("hugeInt").seqInt.gzip
+
+    timed:
+        Set: ^Sequence("hugeFloat") = hugeFloat.gzip
+        assert hugeFloat == Get ^Sequence("hugeFloat").seqFloat.gzip
+
+    timed:
+        Set: ^Sequence("hugeBool") = hugeBool.gzip
+        assert hugeBool == Get ^Sequence("hugeBool").seqBool.gzip
+
+proc testHugeSeqZlib() =
+    Kill ^Sequence
+    timed:
+        Set: ^Sequence("hugeStr") = hugeStr.zlib
+        assert hugeStr == Get ^Sequence("hugeStr").seqString.zlib
+
+    timed:
+        Set: ^Sequence("hugeInt") = hugeInt.zlib
+        assert hugeInt == Get ^Sequence("hugeInt").seqInt.zlib
+
+    timed:
+        Set: ^Sequence("hugeFloat") = hugeFloat.zlib
+        assert hugeFloat == Get ^Sequence("hugeFloat").seqFloat.zlib
+
+    timed:
+        Set: ^Sequence("hugeBool") = hugeBool.zlib
+        assert hugeBool == Get ^Sequence("hugeBool").seqBool.zlib
 
 
 proc testRedirection() =
     var global = "^Sequence"
     Set: @global(5) = join(strSeq,",")
-    assert strSeq == Get @global(5).seqStr
+    assert strSeq == Get @global(5).seqString
 
     let id = 5.5
     Set: @global(id) = join(strSeq,",")
-    assert strSeq == Get @global(id).seqStr
+    assert strSeq == Get @global(id).seqString
 
     global = "^Sequence(6)"
     Set: @global = join(strSeq,",")
-    assert strSeq == Get @global.seqStr
+    assert strSeq == Get @global.seqString
 
 
 when isMainModule:
@@ -102,6 +127,5 @@ when isMainModule:
     test "redirection": testRedirection()
     test "save seq direct": testSeq()
     test "Huge Sequence": testHugeSeq()
-  
-  #dumpData()
-  #dumpKeys()
+    test "Huge Sequence GZIP": testHugeSeqGzip()
+    test "Huge Sequence ZLIB": testHugeSeqZlib()